guix: Guix 1.0.1 -> 1.0.2-0.1.20190917
Upstream still calls this version 1.0.1, but it seems that the preferred
way to package unreleased snapshots in RPM is to use the next version
number. The commit hash used here is obtained from the source code of
guix itself. The reason to update it to a git snapshot is to fix test
failure and 'guix pull' command, which has been broken for months.

https://git.savannah.gnu.org/cgit/guix.git/tree/gnu/packages/package-management.scm?id=f8ddfb505a51a2c794531e090dd7ee97022d1096#n121