arm-linux-gnueabi-toolchain: Use versioned build directory
We should always include the version number in the name of the build
directory, especially when when performing out-of-source builds, to
prevent builds of different versions to use the same directory.

For GCC, LTO is added to the list of enabled languages to be consistent
with the native build of GCC provided by Fedora.

For GLIBC, LTO is removed from compiler flags. We have to do so because
Fedora 33 and later versions enable LTO by default. LTO causes GLIBC to
fail to build and GLIBC isn't known not to support LTO.

The lists of removed flags become long now. Split them into more lines
so they are easier to read.