guix: Drop unused packages
GnuTLS was necessary because Fedora 30 and older versions build GnuTLS
with Guile 2.0, but Guix requires Guile 2.2. Since Fedora 30 is already
EOL, we can delete our patched GnuTLS package.

emacs-graphql was a dependency of emacs-ghub, but emacs-ghub has
switched to use its own GraphQL implementation. Delete it because it is
no longer used by any package in the repository.